+++ spec/whois/record/parser/responses/whois.markmonitor.com/status_registered_spec.rb
@@ -121,13 +121,13 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
       @parser.nameservers.should be_a(Array)
       @parser.nameservers.should have(4).items
       @parser.nameservers[0].should be_a(Whois::Record::Nameserver)
-      @parser.nameservers[0].should == _nameserver.new(:name => "ns2.google.com")
+      @parser.nameservers[0].name.should == "ns2.google.com"
       @parser.nameservers[1].should be_a(Whois::Record::Nameserver)
-      @parser.nameservers[1].should == _nameserver.new(:name => "ns1.google.com")
+      @parser.nameservers[1].name.should == "ns1.google.com"
       @parser.nameservers[2].should be_a(Whois::Record::Nameserver)
-      @parser.nameservers[2].should == _nameserver.new(:name => "ns4.google.com")
+      @parser.nameservers[2].name.should == "ns4.google.com"
       @parser.nameservers[3].should be_a(Whois::Record::Nameserver)
-      @parser.nameservers[3].should == _nameserver.new(:name => "ns3.google.com")
+      @parser.nameservers[3].name.should == "ns3.google.com"
     end
   end
 end